Move unit tests to their own directory. Alter package.json unittest scripts to point to the new location (and allow subdirs).
